Create Plotting the ellipsoid fit to the data
Create a script to scatter plot the first 2 dimensions of the PCA linear vector space, and display a contour representing the boundary, in the first two dimensions, of the the PCA ellipsoid with 2 standard deviations width. 
The steps are:
Use the 'pca' function to find D_pca, the coordinates of the dataset in the PCA vector space, and eigvals, the corresponding eigenvalues that define the width of the axes of the PCA ellipsoid. 
Define the x and y coordinates of the first two dimensions of the data in the PCA vector space, D_pca_x and D_pca_y. Scatter plot this x-y pair using asterisks.
Define x and y coordinates for the shape of the PCA ellipsoid in just the first two dimensions, which is a standard ellipse, as ellipse_x and ellipse_y. The length of the long and short axes radii should be 2 standard deviation distance along each axis. theta=0 should correspond to the positive end of the major axis of the ellipse, with radii equal to 2 times the square root of eigenvalue 1, and theta = pi/2 corresponds to the positive end of the minor axis of the ellipse, with radii 2 times the square root of eigenvalue 2. Plot the ellipse using red dashed lines.