Create Edge strength
Sobel filters are useful for image edge detection. Horizontal and vertical Sobel filters provide positive filter response for similarly oriented bright-to-dark edge transitions and negative response for similarly oriented dark-to-bright edge transitions. Often, we only want to know the magnitude of the edge strength for an edge of any orientation at each pixel.
Write a function called SobelMagnitude, which takes as input an image img, and outputs edge_mag, the magnitude of Sobel filter responses. The magnitude can be found by computing the square root of the sum of the squared horizontal and vertical filter response results.