Create my_fitpca
Create the my_fitpca function that we created in the lecture. It takes as input the [MxN] data matrix D, containing N feature measurements for M samples, and the known classification for each sample in the [Mx1] vector class. The function outputs the struct mdl, which contains the pca models created for each unique classification in class.